Improve Comments
It looks like the current comments in this task are stale, since the flow ownership is now within Pipelines/Templates/update-solution-component-ownership.yml and contains the exact same comments.  

I've tried to update the comments with what I think the purpose of the powershell task is.  btw these yaml pipelines are *awesome*